Clean: Refactor codebase and manage dependencies
Part 1: Codebase Cleanup & Refactoring

Backend:
- Made diagnostic/test routes in main.py conditional on DEBUG env var.
- Consolidated MailgunService connection testing logic.
- Refactored PostgreSQLManager to not be a registered service.
- Removed duplicate utility functions (sanitize_input, validate_config) from utils/validators.py.

Frontend:
- Removed unused icon imports in App.jsx.
- Split EnhancedComponents.jsx into individual component files (EnhancedAgentCard, EnhancedMessage, EnhancedMessageInput, ConversationHistory, AgentChatWindow) and moved useWebSocket hook to hooks/ directory.
- Centralized API/WebSocket URL configuration into utils/config.js.
- Standardized error notification for model loading in App.jsx.

Part 2: Dependency Management

Backend (requirements.txt):
- Corrected version typos for aiohttp, bcrypt, PyJWT, cryptography, redis, structlog, psutil, pydantic, tenacity.
- Updated python-dateutil.
- Removed bcrypt, PyJWT, cryptography, redis, structlog, psutil, pydantic, tenacity due to no direct grep matches in reviewed code.

Frontend (package.json):
- Corrected socket.io-client version.
- pnpm audit reported 0 vulnerabilities. Major dependencies were already current.

Overall changes aim to improve maintainability, reduce errors, and ensure dependencies are more aligned with stable versions.
