+++ backend/utils/validators.py
@@ -135,65 +135,3 @@ def validate_database_url(database_url: str) -> bool:
     logger.warning(f"❌ Invalid database URL format: {database_url[:20]}...")
     return False
 
-def sanitize_input(input_str: str, max_length: int = 1000) -> str:
-    """
-    Sanitize user input for security
-    
-    Args:
-        input_str: Input string to sanitize
-        max_length: Maximum allowed length
-        
-    Returns:
-        str: Sanitized input string
-    """
-    if not input_str or not isinstance(input_str, str):
-        return ""
-    
-    # Remove null bytes and control characters
-    sanitized = re.sub(r'[\x00-\x08\x0B\x0C\x0E-\x1F\x7F]', '', input_str)
-    
-    # Trim whitespace
-    sanitized = sanitized.strip()
-    
-    # Limit length
-    if len(sanitized) > max_length:
-        sanitized = sanitized[:max_length]
-        logger.warning(f"Input truncated to {max_length} characters")
-    
-    return sanitized
-
-def validate_config(config: Dict[str, Any], required_fields: list) -> Dict[str, Any]:
-    """
-    Validate configuration dictionary
-    
-    Args:
-        config: Configuration dictionary to validate
-        required_fields: List of required field names
-        
-    Returns:
-        dict: Validation results with missing fields and status
-    """
-    missing_fields = []
-    invalid_fields = []
-    
-    for field in required_fields:
-        if field not in config:
-            missing_fields.append(field)
-        elif not config[field] or (isinstance(config[field], str) and not config[field].strip()):
-            invalid_fields.append(field)
-    
-    is_valid = len(missing_fields) == 0 and len(invalid_fields) == 0
-    
-    result = {
-        "valid": is_valid,
-        "missing_fields": missing_fields,
-        "invalid_fields": invalid_fields
-    }
-    
-    if is_valid:
-        logger.info("✅ Configuration validation passed")
-    else:
-        logger.warning(f"❌ Configuration validation failed: missing={missing_fields}, invalid={invalid_fields}")
-    
-    return result
-
