+++ frontend/package.json
@@ -54,7 +54,7 @@
     "react-resizable-panels": "^3.0.2",
     "react-router-dom": "^7.6.1",
     "recharts": "^2.15.3",
-    "socket.io-client": "^4.8.1",
+    "socket.io-client": "^4.7.5",
     "sonner": "^2.0.3",
     "tailwind-merge": "^3.3.0",
     "tailwindcss": "^4.1.7",
