+++ requirements.txt
@@ -23,25 +23,25 @@ python-socketio==5.10.0
 
 # Async support
 eventlet==0.33.3
-aiohttp==3.12.13
+aiohttp==3.9.5
 aiofiles==24.1.0
 
 # Additional dependencies
-bcrypt==4.3.0
-PyJWT==2.10.1
-cryptography==45.0.4
-redis==6.2.0
+# bcrypt==4.1.3 # Removed, no direct usage found
+# PyJWT==2.8.0 # Removed, no direct usage found
+# cryptography==43.0.0 # Removed, no direct usage found
+# redis==5.0.7 # Removed, no direct usage found
 email-validator==2.2.0
-structlog==25.4.0
-psutil==7.0.0
+# structlog==24.2.0 # Removed, no direct usage found (standard logging is used)
+# psutil==5.9.8 # Removed, no direct usage found
 
 # Data validation (updated for Python 3.13 compatibility)
-pydantic==2.10.3
+# pydantic==2.7.4 # Removed, no direct usage found
 
 # Utilities
 click==8.1.7
-python-dateutil==2.8.2
-tenacity==9.0.0
+python-dateutil==2.9.0.post0
+# tenacity==8.5.0 # Removed, no direct usage found
 
 # Development dependencies (optional)
 pytest==7.4.3
